requirements.txt
easyocr==1.7.1
loguru==0.7.2
numpy==1.23.5
opencv_python==4.9.0.80
paddleocr==2.7.3
piexif==1.1.3
Pillow==10.3.0
PyQt6==6.7.0
pywin32==306
opencv_contrib_python==4.9.0.80
